JENKINS-53911: Duplicate entries in list, if git repo is checked out twice. Refactor tests
