+++ src/test/java/net/uaznia/lukanus/hudson/plugins/gitparameter/Constants.java
@@ -1,0 +1,12 @@
+package net.uaznia.lukanus.hudson.plugins.gitparameter;
+
+public class Constants {
+    static final String GIT_PARAMETER_REPOSITORY_URL = "https://github.com/jenkinsci/git-parameter-plugin.git";
+    static final String GIT_CLIENT_REPOSITORY_URL = "https://github.com/jenkinsci/git-client-plugin.git";
+    static final String EXAMPLE_REPOSITORY_A_URL = "https://github.com/klimas7/exampleA.git";
+    static final String EXAMPLE_REPOSITORY_B_URL = "https://github.com/klimas7/exampleB.git";
+
+    static final String NAME = "name";
+    static final String PT_REVISION = "PT_REVISION";
+    static final String DEFAULT_VALUE = "defaultValue";
+}
